rename risc0 module to minimize discoverability
